implement first cost function: distance_from_goal_lane
- also implement all the necessary methods for trajectory generation
- also implement helper method `getTrajectoryData` to extract useful
 data from trajectories

+++ behavior_planning/src/snapshot.cpp
@@ -1,6 +1,6 @@
 #include "snapshot.h"
 
-Snapshot::Snapshot(int lane, int s, double v, double a, const string &state)
+Snapshot::Snapshot(int lane, int s, int v, int a, const string &state)
         : lane(lane),
           s(s),
           v(v),
